Separated submission form component.
